agent_gold/data_collector/requirements.txt
pandas
yfinance
requests
openai
tqdm
python-dotenv
pygooglenews
aiohttp